pyproject.toml
# SPDX-FileCopyrightText: 2025 Koen van Greevenbroek
#
# SPDX-License-Identifier: CC-BY-4.0

[project]
name = "food-opt"
version = "0.1.0"
description = "Global food systems optimization model combining environmental and nutritional needs"
authors = [
    {name = "Koen van Greevenbroek", email = "koenvg@stanford.com"}
]
license = "GPL-3.0-or-later"
requires-python = ">=3.12"
dependencies = [
    "snakemake>=9.0",
    "linopy@git+https://github.com/coroa/linopy.git@feat/add-sos-constraints",
    "pypsa@git+https://github.com/koen-vg/PyPSA.git@food-opt",
    "seaborn",
    "numpy",
    "pandas>=2.1",
    "matplotlib",
    "cartopy",
    "networkx",
    "geopandas",
    "shapely>=2.1",
    "rasterio",
    "exactextract",
    "gurobipy",
    "highspy",
    "snakemake-storage-plugin-http",
    "pycountry>=24.6.1",
    "scikit-learn>=1.4",
    "faostat",
    "xlrd>=2.0.1",
    "openpyxl>=3.1.5",
    "gsutil>=5.35",
    "ecmwf-datastores-client",
    "requests",
]

[project.optional-dependencies]
dev = [
    "jupyter",
    "ipython",
    "pre-commit",
    "ruff",
    "sphinx>=8.1.3",
    "sphinx-autodoc-typehints>=3.0.1",
    "furo>=2025.9.25",
    "pydot>=3.0.0",
]

[tool.ruff]
target-version = "py39"
line-length = 88

[tool.ruff.lint]
select = [
    "E",    # pycodestyle errors
    "W",    # pycodestyle warnings
    "F",    # pyflakes
    "I",    # isort
    "N",    # pep8-naming
    "UP",   # pyupgrade
    "B",    # flake8-bugbear
    "C4",   # flake8-comprehensions
    "PIE",  # flake8-pie
    "SIM",  # flake8-simplify
    "RUF",  # Ruff-specific rules
]
ignore = [
    "E501",  # line too long (handled by line-length)
    "B008",  # do not perform function calls in argument defaults
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py files

[tool.ruff.lint.isort]
known-first-party = ["workflow"]
force-sort-within-sections = true

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.uv]
cache-dir = "./.uv-cache"

# [tool.uv.workspace]
# members = [
#     "vendor/PyPSA",
#     "vendor/linopy",
# ]

# [tool.uv.sources]
# pypsa = { workspace = true }
# linopy = { workspace = true }
